+++ SF3000/SFlib/Graphics_Set.h
@@ -1,6 +1,8 @@
 #ifndef __GRAPHICS_SET_DEFINES
 #define __GRAPHICS_SET_DEFINES
 
+#include <stdint.h>
+
 #define EARTH_LAND_BASED_GRAPHICS 0
 #define EARTH_SEA_BASED_GRAPHICS 1
 #define CHEMICAL_GRAPHICS 2
@@ -475,7 +477,7 @@ long test_display_thorg_graphics(long);
 long test_display_space_graphics(long);
 long test_display_death_star_graphics(long);
 
-extern long which_graphics_set;
+extern int32_t which_graphics_set;
 extern char poly_map[128][128];
 extern char collision_map[128][128];
